packet train

+++ net/ipv4/tcp_cubic.c
@@ -430,13 +430,13 @@ static void hystart_update(struct sock *sk, u32 delay)
 				u64 packet_time = now - last_packet_time;
 				if (packet_time > 0 && packet_time <= 2000)
 				{
-					u64 est_bd = (packet_bytes * 8 * 1000000 / packet_time) / (1000 * 1000);
+					u32 est_bd = (packet_bytes * 8 * 1000000 / packet_time) / (1000 * 1000);
 					printk(KERN_INFO "CUBIC (port: %hu) [Round %hu] Now %u, Round Start %u, Bytes ACKed %d, Est. bandwidth %d Mb/s\n",
 						port, round_id, now, ca->round_start, tp->snd_una, est_bd);
 				}
 				else if (packet_time > 2000)
 				{
-					u64 est_bd = (packet_bytes * 8 * 1000000 / packet_time) / (1000 * 1000);
+					u32 est_bd = (packet_bytes * 8 * 1000000 / packet_time) / (1000 * 1000);
 					printk(KERN_INFO "CUBIC (port: %hu) [Round %hu] Now %u, Round Start %u, Bytes ACKed %d, Est. bandwidth %d Mb/s\n",
 						port, round_id, now, ca->round_start, tp->snd_una, est_bd);
 					last_packet_bytes = tp->snd_una;
